Destructor Correction in SinglyLL
In the destructor, we are deleting 'next' pointer, and then pointing 'next to null'. But this is wrong, as the next pointer points to the next node (except the last node that points to NULL), so if we delete 'next' it'll DELETE THE NEXT NODE.

Correction - Point this -> next to NULL directly without deleting 'next' pointer.
IN THE LECTURE VIDEO WE GOT SEGMENTATION ERROR BECAUSE OF IT. which was fixed by manually making 'next' pointer point to NULL before deletion which makes the faulty code to never run (LINE 21) thus no segmentation error, But it's still the wrong code.

I have debugged it properly!